requirements.txt
dnspython==1.16.0
dpkt==1.9.2
NFQP3==0.11
paramiko==2.7.1
pyaes==1.6.1
pyfiglet==0.8.post1
pyscrypt==1.6.2
python-dotenv==0.13.0
requests==2.24.0
RPi.GPIO==0.7.0
scapy==2.4.3.dev691
